Added suuport for Ubuntu linux 64-bit.

+++ provisioners/scripts/common/install_root_user_env.sh
@@ -18,8 +18,14 @@ root_bashrc="${devops_home}/provisioners/scripts/common/users/user-root-bashrc.s
 
 # copy environment profiles to user home.
 cd ${user_home}
-cp -p .bash_profile .bash_profile.orig
-cp -p .bashrc .bashrc.orig
+
+if [ -f ".bash_profile" ]; then
+  cp -p .bash_profile .bash_profile.orig
+fi
+
+if [ -f ".bashrc" ]; then
+  cp -p .bashrc .bashrc.orig
+fi
 
 cp -f ${root_bashprofile} .bash_profile
 cp -f ${root_bashrc} .bashrc
